Working, basic network
This network works! Can choose either MNIST or Chinese callig data.
Both batching and shuffling as we use them mess up our data, as seen by the display function. We need to use TFRecords for these to work.
Need ~15 iters/epochs for 80% accuracy with MNIST, ~400 epochs for 70% accuracy with Chinese.